PRODUCTION_READY/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
lxml
sqlalchemy==2.0.23
